+++ examples/ai-core/src/generate-image/luma-style-reference.ts
@@ -1,5 +1,5 @@
-import { luma } from '@ai-sdk/luma';
-import { experimental_generateImage as generateImage } from 'ai';
+import { luma, LumaImageProviderOptions } from '@ai-sdk/luma';
+import { generateImage } from 'ai';
 import { presentImages } from '../lib/present-image';
 import 'dotenv/config';
 
@@ -10,14 +10,9 @@ async function main() {
     aspectRatio: '1:1',
     providerOptions: {
       luma: {
-        // https://docs.lumalabs.ai/docs/image-generation#style-reference
-        style_ref: [
-          {
-            url: 'https://hebbkx1anhila5yf.public.blob.vercel-storage.com/future-me-8hcBWcZOkbE53q3gshhEm16S87qDpF.jpeg',
-            weight: 0.8,
-          },
-        ],
-      },
+        referenceType: 'style',
+        images: [{ weight: 0.8 }],
+      } satisfies LumaImageProviderOptions,
     },
   });