+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,86 @@
 
 ### Patch Changes
 
+- ai@6.0.0-beta.168
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [475189e]
+  - ai@6.0.0-beta.167
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [9f20c87]
+  - ai@6.0.0-beta.166
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [2625a04]
+  - ai@6.0.0-beta.165
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [cbf52cd]
+  - ai@6.0.0-beta.164
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.163
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [50b70d6]
+  - ai@6.0.0-beta.162
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.161
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [9061dc0]
+  - ai@6.0.0-beta.160
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [3071620]
+  - ai@6.0.0-beta.159
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.158
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [763d04a]
+  - ai@6.0.0-beta.157
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [2406576]
   - ai@6.0.0-beta.156